feat: store user_id and session_id in context
