+++ api/pybot/context.py
@@ -1,0 +1,5 @@
+from contextvars import ContextVar
+
+session_id = ContextVar("session_id", default=None)  # principal:conv_id
+principal = ContextVar("principal", default=None)
+conv_id = ContextVar("conv_id", default=None)
